requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
scrapy==2.11.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.25.2
pandas==2.1.3
python-dotenv==1.0.0
schedule==1.2.0
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
apify-client==1.7.1